Factor out SignInStep and SetupDeviceStep into their own testable modules.
